+++ nequip/data/AtomicData.py
@@ -157,6 +157,13 @@ def _process_dict(kwargs, ignore_fields=[]):
             # ^ this tensor is a scalar; we need to give it
             # a data dimension to play nice with irreps
             kwargs[k] = v
+        elif isinstance(v, torch.Tensor):
+            # This is a tensor, so we just don't do anything except avoid the warning in the `else`
+            pass
+        else:
+            warnings.warn(
+                f"Value for field {k} was of unsupported type {type(v)} (value was {v})"
+            )
 
     if AtomicDataDict.BATCH_KEY in kwargs:
         num_frames = kwargs[AtomicDataDict.BATCH_KEY].max() + 1
@@ -229,7 +236,6 @@ class AtomicData(Data):
     def __init__(
         self, irreps: Dict[str, e3nn.o3.Irreps] = {}, _validate: bool = True, **kwargs
     ):
-
         # empty init needed by get_example
         if len(kwargs) == 0 and len(irreps) == 0:
             super().__init__()
@@ -420,7 +426,6 @@ class AtomicData(Data):
         )
 
         if atoms.calc is not None:
-
             if isinstance(
                 atoms.calc, (SinglePointCalculator, SinglePointDFTCalculator)
             ):